+++ src/tests/runtime_code.rs
@@ -30,6 +30,7 @@ contract Test {
     super::build_solidity(
         sources,
         BTreeMap::new(),
+        None,
         SolcPipeline::Yul,
         compiler_llvm_context::OptimizerSettings::cycles(),
     )