+++ src/yul/parser/error.rs
@@ -2,6 +2,8 @@
 //! The Yul IR parser error.
 //!
 
+use std::collections::BTreeSet;
+
 use crate::yul::lexer::token::location::Location;
 
 ///
@@ -50,5 +52,13 @@ pub enum Error {
         expected: String,
         /// The invalid identifier.
         found: String,
+    },
+    /// Invalid attributes.
+    #[error("{location} Found invalid LLVM attributes: {values:?}")]
+    InvalidAttributes {
+        /// The invalid token location.
+        location: Location,
+        /// The list of invalid attributes.
+        values: BTreeSet<String>,
     },
 }