+++ src/yul/parser/statement/expression/function_call/name.rs
@@ -309,7 +309,7 @@ impl Name {
     ///
     fn parse_verbatim(input: &str) -> Option<Self> {
         let verbatim = input.strip_prefix("verbatim")?;
-        let regex = regex::Regex::new(r#"_(\d+)i_(\d+)o"#).expect("Always valid");
+        let regex = regex::Regex::new(r"_(\d+)i_(\d+)o").expect("Always valid");
         let captures = regex.captures(verbatim)?;
         let input_size: usize = captures.get(1)?.as_str().parse().ok()?;
         let output_size: usize = captures.get(2)?.as_str().parse().ok()?;