+++ src/yul/parser/statement/function_definition.rs
@@ -2,6 +2,7 @@
 //! The function definition statement.
 //!
 
+use std::collections::BTreeSet;
 use std::collections::HashSet;
 
 use inkwell::types::BasicType;
@@ -38,9 +39,17 @@ pub struct FunctionDefinition {
     pub result: Vec<Identifier>,
     /// The function body block.
     pub body: Block,
+    /// The function LLVM attributes encoded in the identifier.
+    pub attributes: BTreeSet<compiler_llvm_context::EraVMAttribute>,
 }
 
 impl FunctionDefinition {
+    /// The LLVM attribute section prefix.
+    pub const LLVM_ATTRIBUTE_PREFIX: &'static str = "$llvm_";
+
+    /// The LLVM attribute section suffix.
+    pub const LLVM_ATTRIBUTE_SUFFIX: &'static str = "_llvm$";
+
     ///
     /// The element parser.
     ///
@@ -62,6 +71,7 @@ impl FunctionDefinition {
                 .into());
             }
         };
+        let identifier = Identifier::new(location, identifier.inner);
 
         match FunctionName::from(identifier.inner.as_str()) {
             FunctionName::UserDefined(_) => {}
@@ -159,20 +169,62 @@ impl FunctionDefinition {
 
         let body = Block::parse(lexer, next)?;
 
+        let attributes = Self::get_llvm_attributes(&identifier)?;
+
         Ok(Self {
             location,
             identifier: identifier.inner,
             arguments,
             result,
             body,
+            attributes,
         })
     }
 
     ///
-    /// Get the list of missing deployable libraries.
+    /// Gets the list of missing deployable libraries.
     ///
     pub fn get_missing_libraries(&self) -> HashSet<String> {
         self.body.get_missing_libraries()
+    }
+
+    ///
+    /// Gets the list of LLVM attributes provided in the function name.
+    ///
+    pub fn get_llvm_attributes(
+        identifier: &Identifier,
+    ) -> Result<BTreeSet<compiler_llvm_context::EraVMAttribute>, Error> {
+        let mut valid_attributes = BTreeSet::new();
+
+        let llvm_begin = identifier.inner.find(Self::LLVM_ATTRIBUTE_PREFIX);
+        let llvm_end = identifier.inner.find(Self::LLVM_ATTRIBUTE_SUFFIX);
+        let attribute_string = if let (Some(llvm_begin), Some(llvm_end)) = (llvm_begin, llvm_end) {
+            if llvm_begin < llvm_end {
+                &identifier.inner[llvm_begin + Self::LLVM_ATTRIBUTE_PREFIX.len()..llvm_end]
+            } else {
+                return Ok(valid_attributes);
+            }
+        } else {
+            return Ok(valid_attributes);
+        };
+
+        let mut invalid_attributes = BTreeSet::new();
+        for value in attribute_string.split('_') {
+            match compiler_llvm_context::EraVMAttribute::try_from(value) {
+                Ok(attribute) => valid_attributes.insert(attribute),
+                Err(value) => invalid_attributes.insert(value),
+            };
+        }
+
+        if !invalid_attributes.is_empty() {
+            return Err(ParserError::InvalidAttributes {
+                location: identifier.location,
+                values: invalid_attributes,
+            }
+            .into());
+        }
+
+        Ok(valid_attributes)
     }
 }
 
@@ -206,6 +258,11 @@ where
             self.result.len(),
             Some(inkwell::module::Linkage::Private),
         )?;
+        compiler_llvm_context::EraVMFunction::set_function_attributes(
+            context.llvm(),
+            function.borrow().declaration(),
+            self.attributes.clone().into_iter().collect(),
+        );
         function
             .borrow_mut()
             .set_yul_data(compiler_llvm_context::EraVMFunctionYulData::default());
@@ -326,6 +383,8 @@ where
 
 #[cfg(test)]
 mod tests {
+    use std::collections::BTreeSet;
+
     use crate::yul::lexer::token::location::Location;
     use crate::yul::lexer::Lexer;
     use crate::yul::parser::error::Error;
@@ -579,6 +638,81 @@ object "Test" {
             Err(Error::ReservedIdentifier {
                 location: Location::new(14, 22),
                 identifier: "basefee".to_owned()
+            }
+            .into())
+        );
+    }
+
+    #[test]
+    fn error_invalid_attributes_single() {
+        let input = r#"
+object "Test" {
+    code {
+        {
+            return(0, 0)
+        }
+    }
+    object "Test_deployed" {
+        code {
+            {
+                return(0, 0)
+            }
+
+            function test_$llvm_UnknownAttribute_llvm$_test() -> result {
+                result := 42
+            }
+        }
+    }
+}
+    "#;
+        let mut invalid_attributes = BTreeSet::new();
+        invalid_attributes.insert("UnknownAttribute".to_owned());
+
+        let mut lexer = Lexer::new(input.to_owned());
+        let result = Object::parse(&mut lexer, None);
+        assert_eq!(
+            result,
+            Err(Error::InvalidAttributes {
+                location: Location::new(14, 22),
+                values: invalid_attributes,
+            }
+            .into())
+        );
+    }
+
+    #[test]
+    fn error_invalid_attributes_multiple_repeated() {
+        let input = r#"
+object "Test" {
+    code {
+        {
+            return(0, 0)
+        }
+    }
+    object "Test_deployed" {
+        code {
+            {
+                return(0, 0)
+            }
+
+            function test_$llvm_UnknownAttribute1_UnknownAttribute1_UnknownAttribute2_llvm$_test() -> result {
+                result := 42
+            }
+        }
+    }
+}
+    "#;
+        let mut invalid_attributes = BTreeSet::new();
+        invalid_attributes.insert("UnknownAttribute1".to_owned());
+        invalid_attributes.insert("UnknownAttribute2".to_owned());
+
+        let mut lexer = Lexer::new(input.to_owned());
+        let result = Object::parse(&mut lexer, None);
+        assert_eq!(
+            result,
+            Err(Error::InvalidAttributes {
+                location: Location::new(14, 22),
+                values: invalid_attributes,
             }
             .into())
         );